gcp/cloudrun/gui/requirements.txt
Flask
gunicorn
google-cloud-firestore
requests